# docs/OpenAPI/frontend-api.yaml
openapi: 3.0.0
info:
  title: frontend
  version: '1.0'
  contact: {}
  description: Public facing application REST API
servers:
  - url: 'http://localhost:8888'
paths:
  /userInfo:
    get:
      summary: Get User Info
      tags:
        - workout-app
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  FirstName:
                    type: string
                  LastName:
                    type: string
                  Weight:
                    type: number
                  WaistCirc:
                    type: number
                  HeightInches:
                    type: number
                  LeanBodyMass:
                    type: number
                  Age:
                    type: number
                  Gender:
                    type: string
              examples:
                example:
                  value:
                    FirstName: Anthony
                    LastName: Hannah
                    Weight: 215
                    WaistCirc: 11
                    HeightInches: 72
                    LeanBodyMass: 15
                    Age: 27
                    Gender: female
        '401':
          description: Unauthorized
        '428':
          description: Precondition Required - missing session-token
      operationId: get-userInfo
      description: Returns all user profile data
      parameters:
        - schema:
            type: string
          in: header
          name: Session-Token
          description: The session token of the currently logged in user
          required: true
    parameters: []
    post:
      summary: Update User Profile
      operationId: post-userProfile
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '428':
          description: Bad Request - missing session-token
      parameters:
        - schema:
            type: string
          in: header
          name: Session-Token
          description: The session token of the currently logged in user
          required: true
      description: Updates user profile data
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                FirstName:
                  type: string
                LastName:
                  type: string
                Weight:
                  type: number
                WaistCirc:
                  type: number
                HeightInches:
                  type: number
                LeanBodyMass:
                  type: number
                Age:
                  type: number
                Gender:
                  type: string
      tags:
        - workout-app
  /apiVersion:
    get:
      summary: Get API Version
      tags:
        - workout-app
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiVersion:
                    type: number
      operationId: get-apiVersion
      description: Returns the symantec version number of the api
    parameters: []
  /userBaseline:
    parameters: []
    post:
      summary: Update User Baseline
      operationId: post-userBaseline
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '428':
          description: Precondition Required - missing session-token
      parameters:
        - schema:
            type: string
          in: header
          name: Session-Token
          description: The session token of the currently logged in user
          required: true
      description: Updates user baseline information
      tags:
        - workout-app
  '/userWeeklyTracking/{week}':
    parameters:
      - schema:
          type: string
        name: week
        in: path
        required: true
    post:
      summary: Post User Weekly Tracking
      operationId: post-userWeeklyTracking
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '428':
          description: Precondition Required - missing session-token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Session-Token:
                  type: string
                  description: The session token of the currently logged in user
              required:
                - Session-Token
        description: 'TODO: finish userWeeklyTracking-post'
      tags:
        - workout-app
      description: Updates single week in weekly tracking
      parameters:
        - schema:
            type: string
          in: header
          name: Session-Token
          description: The session token of the currently logged in user
          required: true
components:
  schemas: {}
tags:
  - name: workout-app